Move iOS component rule to base plugin
